rencongAI-hafiz/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
langchain
langchain_community
langchain_google_genai
langchain_huggingface
sentence-transformers
pandas
protobuf
pydantic
python-dotenv
requests
slowapi
streamlit
torch
transformers
faiss-cpu